+++ src/renderer/src/components/MyDrawer.tsx
@@ -2,88 +2,16 @@ import * as React from 'react';
 import {useTheme} from '@mui/material/styles';
 import Drawer from '@mui/material/Drawer';
 import CssBaseline from '@mui/material/CssBaseline';
-import List from '@mui/material/List';
 import Divider from '@mui/material/Divider';
-import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import BestBoutArea from "./Areas/BestBoutArea";
-import {For, HTMLChakraProps, Text, Box, IconButton, VStack, IconButtonProps} from "@chakra-ui/react";
-import TimerOutlinedIcon from '@mui/icons-material/TimerOutlined';
-import EditNoteIcon from '@mui/icons-material/EditNote';
-import ChatBubbleIcon from '@mui/icons-material/ChatBubble';
-import AddTaskIcon from '@mui/icons-material/AddTask';
-import {Tweet} from "./Dialogs/Tweet";
-import {StopWatch} from "./Dialogs/StopWatch";
-import {AddLog} from "./Dialogs/AddLog";
-import {AddTask} from "./Dialogs/AddTask";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
+import {HTMLChakraProps, Text, Box, IconButton} from "@chakra-ui/react";
+import {CommandList} from "./SideBar/CommandList";
+import {MainViewer} from "./SideBar/MainViewer";
+import {CommandBar} from "./SideBar/CommandBar";
 
 const drawerWidth = 280;
-
-interface MainProps extends HTMLChakraProps<"div">{
-    open?: boolean;
-    drawerWidth: number;
-}
-
-const Main: React.FC<MainProps> = ({ open = false, drawerWidth, children,...other }) => {
-    const theme = useTheme();
-    const marginLeft = open ? 0 : `-${drawerWidth}px`;
-
-    const transition = open
-        ? `${theme.transitions.duration.enteringScreen}ms ${theme.transitions.easing.easeOut}`
-        : `${theme.transitions.duration.leavingScreen}ms ${theme.transitions.easing.sharp}`;
-
-    return (
-        <Box
-            flexGrow={1}
-            h="100vh"
-            transition={`margin ${transition}`}
-            ml={marginLeft}
-            pl={open?2:0}
-            overflowY={"hidden"}
-            {...other}
-        >
-            {children}
-        </Box>
-    );
-};
-
-interface AppBarProps extends HTMLChakraProps<"div">{
-    open?: boolean;
-    drawerWidth: number;
-}
-
-const AppBar: React.FC<AppBarProps&{isVisible:boolean}> = ({ open = false, drawerWidth, children,isVisible=false }) => {
-    const theme = useTheme();
-
-    // 動的にスタイルを決定
-    const marginLeft = open ? `${drawerWidth}px` : "0";
-    const transition = open
-        ? `${theme.transitions.duration.enteringScreen}ms ${theme.transitions.easing.easeOut}`
-        : `${theme.transitions.duration.leavingScreen}ms ${theme.transitions.easing.sharp}`;
-
-    return (
-        <Box
-            as="header"
-            bg="transparent"
-            color="#000000"
-            boxShadow="none"
-            transition={transition}
-            width={"55px"}
-            minHeight="100vh"
-            height="100vh"
-            marginLeft={marginLeft}
-            display={!open?"flex":"none"}
-            justifyContent="flex-end"
-        >
-            { isVisible?children:<></>}
-        </Box>
-    );
-};
 
 function DrawerHeader({ children,...other }:HTMLChakraProps<"div">) {
     return (
@@ -91,7 +19,7 @@ function DrawerHeader({ children,...other }:HTMLChakraProps<"div">) {
             display="flex"
             alignItems="center"
             px={1} // 左右のパディング
-            py={0} // 上下のパディング
+            py={3} // 上下のパディング
             justifyContent="space-between"
             {...other}
         >
@@ -100,11 +28,10 @@ function DrawerHeader({ children,...other }:HTMLChakraProps<"div">) {
     );
 }
 
-export default function SideBarFrame({children}:{children:React.ReactNode}) {
+export function SideBar({children}:{children:React.ReactNode}) {
     const theme = useTheme();
     const [open, setOpen] = React.useState(true);
-    const [isAppBarVisible, setIsAppBarVisible] = React.useState(false);
-    // const {windowHeight} = useWindowSize();
+    const [isCommandBarVisible, setIsCommandBarVisible] = React.useState(false);
 
     const handleDrawerOpen = () => {
         setOpen(true);
@@ -117,71 +44,25 @@ export default function SideBarFrame({children}:{children:React.ReactNode}) {
     React.useEffect(() => {
         // Drawerのopen/closeとAppBarの表示をずらす
         if (open) {
-            setIsAppBarVisible(false);
+            setIsCommandBarVisible(false);
             return ()=>{};
         } else {
             const timer = setTimeout(() => {
-                setIsAppBarVisible(true);
+                setIsCommandBarVisible(true);
             },200);
             return () => clearTimeout(timer); // クリーンアップ
         }
     },[open]);
-    const sideBarIconButtonStyle :IconButtonProps ={
-        size:"xl",
-        textStyle:"xl",
-        color:"orange.900",
-        bg:{
-            base: "transparent",
-            _hover:"blackAlpha.200"
-        }
-    }
 
     return (
         <Box display={"flex"}>
             <CssBaseline />
-            <AppBar open={open} isVisible={isAppBarVisible} drawerWidth={drawerWidth}>
-                <VStack>
-                    <IconButton
-                        {...sideBarIconButtonStyle}
-                        onClick={handleDrawerOpen}
-                        display={open?'none':"block"}
-                        variant={"subtle"}
-                    >
-                        <MenuIcon />
-                    </IconButton>
-
-                    <Divider />
-
-                    {/*サイドバー*/}
-                    {/*ストップウォッチ*/}
-                    <StopWatch>
-                        <IconButton {...sideBarIconButtonStyle}>
-                            <TimerOutlinedIcon/>
-                        </IconButton>
-                    </StopWatch>
-
-                    {/*記録*/}
-                    <AddLog>
-                        <IconButton {...sideBarIconButtonStyle}>
-                            <EditNoteIcon/>
-                        </IconButton>
-                    </AddLog>
-
-                    {/*タスクの追加*/}
-                    <AddTask>
-                        <IconButton {...sideBarIconButtonStyle}>
-                            <ChatBubbleIcon/>
-                        </IconButton>
-                    </AddTask>
-
-                    {/*つぶやき*/}
-                    <Tweet>
-                        <IconButton {...sideBarIconButtonStyle}>
-                            <AddTaskIcon/>
-                        </IconButton>
-                    </Tweet>
-                </VStack>
-            </AppBar>
+            <CommandBar
+                open={open}
+                drawerWidth={drawerWidth}
+                handleDrawerOpen={handleDrawerOpen}
+                isVisible={isCommandBarVisible}
+            />
             <Drawer
                 sx={{
                     width: drawerWidth,
@@ -214,78 +95,28 @@ export default function SideBarFrame({children}:{children:React.ReactNode}) {
                     </IconButton>
                 </DrawerHeader>
                 <Divider />
+
                 {/*コマンドエリア*/}
-                <List>
-                    <For
-                        each={[
-                            { name: "ストップウォッチ", icon:<TimerOutlinedIcon/>,dialog:"stopwatch"},
-                            { name: "記録", icon:<EditNoteIcon/>,dialog:"log"},
-                            { name: "つぶやき", icon:<ChatBubbleIcon/>,dialog:"tweet"},
-                            { name: "タスクを追加", icon:<AddTaskIcon/>,dialog:"task"},
-                        ]}
-                    >
-                        {(item) =>{
-                            switch (item.dialog) {
-                                case "stopwatch":
-                                    return (
-                                        <StopWatch key={"StopWatch_List"}>
-                                            <ListButton name={item.name} icon={item.icon}/>
-                                        </StopWatch>
-                                    );
-                                case "log":
-                                    return (
-                                        <AddLog key={"AddLog_List"}>
-                                            <ListButton name={item.name} icon={item.icon}/>
-                                        </AddLog>
-                                    );
-                                case "task":
-                                    return (
-                                        <AddTask key={"AddTask_List"}>
-                                            <ListButton name={item.name} icon={item.icon}/>
-                                        </AddTask>
-                                    );
-                                case "tweet":
-                                    return (
-                                        <Tweet key={"AddTweet_List"}>
-                                            <ListButton name={item.name} icon={item.icon}/>
-                                        </Tweet>
-                                    );
-                                default:
-                                    return (<></>);
-                            }
-                        }}
-                    </For>
-                </List>
+                <CommandList/>
                 <Divider />
+
                 {/*ベストバウトエリア*/}
                 <Box width={"100%"} height={"100%"} padding={2.5} display={"flex"} alignItems={"flex-end"}>
                     <BestBoutArea data={["1月8日","2月20日","3月10日"]}/>
                 </Box>
             </Drawer>
-            <Main
+            <MainViewer
                 open={open}
                 drawerWidth={drawerWidth}
             >
                 <Box h={"100%"} width={"100%"}>
                     {children}
                 </Box>
-            </Main>
+            </MainViewer>
         </Box>
     );
 }
 
-function ListButton(props:{name:string,icon:React.ReactElement}) {
-    return (
-        <ListItem disablePadding>
-            <ListItemButton>
-                <ListItemIcon>
-                    {props.icon}
-                </ListItemIcon>
-                <ListItemText
-                    disableTypography
-                    primary={<Text fontSize={"1.1em"}>{props.name}</Text>}
-                />
-            </ListItemButton>
-        </ListItem>
-    );
-}
+
+
+export default SideBar;
